options: fix integer overflow in option parser
As the numeric range of CoAP option codes is [0..65804], the data
type for representing these codes must be uint32_t rather than
uint16_t.

Fixes OSS-Fuzz 26371. [1]

[1] https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=26371

+++ include/coap2/option.h
@@ -374,7 +374,7 @@ uint16_t coap_opt_delta(const coap_opt_t *opt);
  *
  * @return     The option's length or @c 0 when undefined.
  */
-uint16_t coap_opt_length(const coap_opt_t *opt);
+uint32_t coap_opt_length(const coap_opt_t *opt);
 
 /**
  * Returns a pointer to the value of the given option. @p opt must point to an
